test bigdecimal

+++ src/main/java/com/example/calcservice/triangle/controller/TriangleController.java
@@ -4,6 +4,8 @@ import com.example.calcservice.triangle.service.TriangleInterface;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @CrossOrigin(origins = {"http://localhost:3000"})
 
 @RestController
@@ -13,12 +15,12 @@ import org.springframework.web.bind.annotation.*;
 public class TriangleController {
     private TriangleInterface triangleInterface;
     @GetMapping
-    public TriangleDTO resultCalcTriangle(@RequestParam(defaultValue = "0")  double a,
-                                          @RequestParam(defaultValue = "0")  double b,
-                                          @RequestParam(defaultValue = "0")  double c,
-                                          @RequestParam(defaultValue = "0")  double angleA,
-                                          @RequestParam(defaultValue = "0")  double angleB,
-                                          @RequestParam(defaultValue = "0")  double angleC)
+    public TriangleDTO resultCalcTriangle(@RequestParam(defaultValue = "0")  BigDecimal a ,
+                                          @RequestParam(defaultValue = "0")  BigDecimal b,
+                                          @RequestParam(defaultValue = "0")  BigDecimal c,
+                                          @RequestParam(defaultValue = "0")  BigDecimal angleA,
+                                          @RequestParam(defaultValue = "0")  BigDecimal angleB,
+                                          @RequestParam(defaultValue = "0")  BigDecimal angleC)
     {
         return triangleInterface.resultCalcTriangle(a, b, c, angleA, angleB, angleC);
     }
