+++ src/main/java/com/example/calcservice/triangle/service/TriangleInterface.java
@@ -2,11 +2,12 @@ package com.example.calcservice.triangle.service;
 import com.example.calcservice.triangle.model.Triangle;
 import com.example.calcservice.triangle.model.TriangleDTO;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 
 public interface TriangleInterface {
-    TriangleDTO resultCalcTriangle(double a, double b, double c,
-                                   double angleA, double angleB, double angleC);
+    TriangleDTO resultCalcTriangle(BigDecimal a, BigDecimal b, BigDecimal c,
+                                   BigDecimal angleA, BigDecimal angleB, BigDecimal angleC);
 
 }
